refactor: split benchmark commands.go into separate files
- types.go: shared result types
- check.go: RunCheck, PrintCheckResult
- compare.go: RunCompare, PrintCompareResult
- lint.go: RunLint, PrintLintResult
- catalog.go: RunCatalog, PrintCatalogResult
- baseline.go: baseline management
- calibrate.go: threshold calibration
- tune.go: weight tuning
- runtime.go: Go benchmark performance
